# swagger-api.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/admin/days-off": {
            "post": {
                "description": "Mark a specific date as unavailable for appointments (holiday, closure, etc.)",
                "operationId": "AdminDaysOffController_create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateDayOffDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Day off created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DayOffResponseDto"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Day off already exists for this date"
                    }
                },
                "summary": "Add a day off",
                "tags": [
                    "Admin - Days Off"
                ]
            },
            "get": {
                "operationId": "AdminDaysOffController_getAll",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "List of all days off",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DayOffResponseDto"
                                    }
                                }
                            }
                        }
                    }
                },
                "summary": "List all days off",
                "tags": [
                    "Admin - Days Off"
                ]
            }
        },
        "/admin/days-off/{id}": {
            "put": {
                "operationId": "AdminDaysOffController_update",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "Day off ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateDayOffDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Day off updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DayOffResponseDto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Day off not found"
                    }
                },
                "summary": "Update a day off",
                "tags": [
                    "Admin - Days Off"
                ]
            },
            "delete": {
                "operationId": "AdminDaysOffController_delete",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "Day off ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Day off removed successfully"
                    },
                    "404": {
                        "description": "Day off not found"
                    }
                },
                "summary": "Remove a day off",
                "tags": [
                    "Admin - Days Off"
                ]
            }
        },
...
    },
    "info": {
        "title": "Appointment Scheduler API",
        "description": "Production-ready appointment scheduling system",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "schemas": {
...
            "CreateDayOffDto": {
                "type": "object",
                "properties": {
                    "date": {
                        "type": "string",
                        "example": "2025-12-25",
                        "description": "Date to mark as day off (YYYY-MM-DD)"
                    },
                    "note": {
                        "type": "string",
                        "example": "Christmas Day",
                        "description": "Optional note about the day off",
                        "maxLength": 255
                    }
                },
                "required": [
                    "date"
                ]
            },
            "DayOffResponseDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "f47ac10b-58cc-4372-a567-0e02b2c3d479"
                    },
                    "date": {
                        "type": "string",
                        "example": "2025-12-25"
                    },
                    "note": {
                        "type": "string",
                        "example": "Christmas Day",
                        "nullable": true
                    },
                    "createdAt": {
                        "format": "date-time",
                        "type": "string",
                        "description": "When this day off was created"
                    }
                },
                "required": [
                    "id",
                    "date",
                    "note",
                    "createdAt"
                ]
            },
            "UpdateDayOffDto": {
                "type": "object",
                "properties": {
                    "note": {
                        "type": "string",
                        "example": "Christmas Day (Updated)",
                        "description": "Update the note about the day off",
                        "maxLength": 255
                    }
                }
            },
...
        }
    }
}